Update Sysmon_OSSEC-Decoders.xml
This one isn't perfect either... data always picks up an extra 'P' - I can't match the last repeating instance of Parent anymore in this decode.  I think there are so many repeating patterns that it is difficult for OSSEC to pull out chunks for the decode.  I feel like I'm hitting a limitation or issue with OSSEC that isn't well-documented.

This has been quite a humbling experience for me on what I thought I knew about OSSEC decoders.

Overall, if you don't care for the change or behavior of this submission.  It won't hurt my feelings if you reject it.  It seems to work with spaces in all sections (username, image name and path and parents).

If you delete the "arent\.*" right before regex closes, you'll capture both the image name and the command line in data.  This might not be a bad option if you're trying to <match> names of executable files.